requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.6.2
defusedxml==0.7.1
fonttools==4.53.0
fpdf2==2.7.9
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pillow==10.3.0
pydantic==2.7.3
pydantic_core==2.18.4
typing_extensions==4.12.1
yarl==1.9.4
